+++ Sources/SwifterSwift/AppKit/NSStackViewExtensions.swift
@@ -1,0 +1,26 @@
+// NSStackViewExtensions.swift - Copyright 2024 SwifterSwift
+
+#if canImport(AppKit) && !targetEnvironment(macCatalyst)
+import AppKit
+
+// MARK: - Methods
+
+public extension NSStackView {
+    /// SwifterSwift: Adds array of views to the end of the arrangedSubviews array.
+    ///
+    /// - Parameter views: views array.
+    func addArrangedSubviews(_ views: [NSView]) {
+        for view in views {
+            addArrangedSubview(view)
+        }
+    }
+
+    /// SwifterSwift: Removes all views in stack’s array of arranged subviews.
+    func removeArrangedSubviews() {
+        for view in arrangedSubviews {
+            removeArrangedSubview(view)
+        }
+    }
+}
+
+#endif